requirements.txt
boto3==1.24.12
botocore==1.27.12
gTTS==2.2.4
moviepy==1.0.3
mutagen==1.45.1
playwright==1.23.0
praw==7.6.0
python-dotenv==0.20.0
pytube==12.1.0
requests==2.28.1
rich==12.4.4
translators==5.2.2
